+++ gen/templates/profile_parcels/resource.go
@@ -426,7 +426,12 @@ func (r *{{camelCase .Name}}ProfileParcelResource) Read(ctx context.Context, req
 		return
 	}
 
-	state.updateFromBody(ctx, res)
+	// If every attribute is set to null we are dealing with an import operation and therefore reading all attributes
+	if state.isNull(ctx, res) {
+		state.fromBody(ctx, res)
+	} else {
+		state.updateFromBody(ctx, res)
+	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Name.ValueString()))